+++ services/madoc-ts/src/extensions/capture-models/Paragraphs/Paragraphs.slots.tsx
@@ -71,12 +71,13 @@ const SingleParagraphEntity: ProfileConfig['SingleEntity'] = ({ showTitle = true
       <Slots.AdjacentNavigation>
         {isModified && <ModifiedStatus />}
         <Slots.InlineSelector />
-        {mapProperties(entity, ({ hasSelector, type, label, description, property, canInlineField }) => {
+        {mapProperties(entity, ({ hasSelector, type, label, description, disableForm, property, canInlineField }) => {
           return (
             <Slots.InlineProperties
               hasSelector={hasSelector}
               type={type}
               property={property}
+              disableForm={disableForm}
               label={label}
               description={description}
               canInlineField={canInlineField}