+++ services/madoc-ts/src/frontend/shared/capture-models/editor/input-types/TaggedTextField/TaggedTextField.tsx
@@ -28,6 +28,7 @@ export interface TaggedTextFieldProps extends BaseField {
   enableLineBreak?: boolean;
   enablePageBreak?: boolean;
   enableIllegible?: boolean;
+  disabled?: boolean;
 }
 
 export type TagDefinition = {