+++ services/madoc-ts/src/frontend/shared/capture-models/helpers/resolve-selector.ts
@@ -1,6 +1,6 @@
 import { BaseSelector } from '../types/selector-types';
 
-export function resolveSelector(selector: BaseSelector, revisionId?: string): BaseSelector {
+export function resolveSelector(selector: BaseSelector, revisionId?: string | null): BaseSelector {
   if (selector.revisedBy && revisionId) {
     for (const revisedSelector of selector.revisedBy) {
       if (revisedSelector.revisionId === revisionId) {