+++ services/madoc-ts/src/frontend/shared/capture-models/plugin-api/hooks/use-field.ts
@@ -5,9 +5,10 @@ import { useFieldPlugin } from './use-field-plugin';
 export function useField<Props extends BaseField>(
   props: Props,
   value: Props['value'],
-  updateValue: (value: Props['value']) => void
+  updateValue: (b: Props['value']) => void,
+  options: { disabled?: boolean } = {}
 ) {
   const field = useFieldPlugin(props.type);
 
-  return createElement(field.Component, { ...props, value, updateValue });
+  return createElement(field.Component, { ...props, value, updateValue, ...options });
 }