+++ services/madoc-ts/src/frontend/shared/form/FieldSet.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const FieldSet = styled.fieldset`
+  border: none;
+  margin: 0;
+  padding: 0;
+`;